connection with map component, draw a path
