+++ src/components/Map/Map.tsx
@@ -7,6 +7,7 @@ import './leaflet.css';
 import './Map.scss';
 import { useEffect, useState } from 'react';
 import { ObjectData } from '../../models';
+import { getPath } from './utils.ts';
 
 function OnMapClick() {
   useMapEvents({
@@ -16,46 +17,6 @@ function OnMapClick() {
   });
 
   return null;
-}
-
-async function getPath(
-  location: L.LatLng,
-  destination: L.LatLng,
-  type: 'foot' | 'bike'
-): Promise<[L.LatLng, L.LatLng][]> {
-  const newRoute: [L.LatLng, L.LatLng][] = [];
-  const routePoints: L.LatLng[] = [];
-
-  // I would suggest maybe placing the logic inside useEffect, using state change setRoute and updating it when location or destination change
-  // useEffect(() => {
-  //   if (location && destination) { logic }
-  // }, [location, destination])
-  // and placing route on the map via:
-  // const [route, setRoute] = useState<[L.LatLng, L.LatLng][]>([])
-  // <MapContainer>
-  //   {route.length > 0 && <Polyline positions={route}/>}
-  // </MapContainer>
-
-  fetch(
-    `https://routing.openstreetmap.de/routed-foot/route/v1/${type}/${location.lng},${location.lat};${destination.lng},${destination.lat}?overview=false&steps=true`
-  )
-    .then((response) => response.json())
-    .then((data) => {
-      data['routes'][0]['legs'].map((legs: any) => {
-        legs['steps'].map((step: any) => {
-          step['intersections'].map((intersection: any) => {
-            const position: [number, number] = intersection['location'];
-            routePoints.push(new L.LatLng(position[1], position[0]));
-          });
-        });
-      });
-      routePoints.slice(0, -1).map((_, i) => {
-        newRoute.push([routePoints[i], routePoints[i + 1]]);
-      });
-      return newRoute;
-      // at this point we have newRoute, we could do setRoute(newRoute) here, then no return
-    })
-    .catch((error) => console.log(error));
 }
 
 export const Map = () => {
@@ -70,17 +31,18 @@ export const Map = () => {
   const wzimCoords = L.latLng(52.16198, 21.04633);
 
   useEffect(() => {
-    getPath(
-      {
-        lat: points?.startPoint.lat,
-        lng: points?.startPoint.lng,
-      } as L.LatLng,
-      {
-        lat: points?.destinationPoint.lat,
-        lng: points?.destinationPoint.lng,
-      } as L.LatLng,
-      'foot'
-    ).then((data) => console.log(data));
+    if (points?.startPoint && points?.destinationPoint)
+      getPath(
+        {
+          lat: points?.startPoint.lat,
+          lng: points?.startPoint.lng,
+        } as L.LatLng,
+        {
+          lat: points?.destinationPoint.lat,
+          lng: points?.destinationPoint.lng,
+        } as L.LatLng,
+        'foot'
+      ).then((data) => setRoad(data));
   }, [points]);
 
   return (
@@ -99,7 +61,7 @@ export const Map = () => {
           attribution='&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors'
         />
         <CustomMarker position={wzimCoords} text={'WZIM'} />
-        {road.length > 0 && <Polyline route={road} />}
+        <Polyline positions={road} />
         <OnMapClick />
       </MapContainer>
     </div>
